media: ar0144: Add PIXEL_RATE and update LINK_FREQ control
The link frequency needs now to be supplied by the fwnode. The MIPI-CSI2
interface requires three link frequencies to be supplied. One for 8, 10
and 12 bit which represent the maximum and target link frequency for
that configuration. The parallel interface requires one frequency which
represents the maximum and target pixel clock.

Add a PIXLE_RATE control which returns the currently configured pixel
clock frequency.

Make sure the PLL is calculated initially so these controls have valid
values to work with.

Signed-off-by: Stefan Riedmueller <s.riedmueller@phytec.de>
Signed-off-by: Christian Hemp <c.hemp@phytec.de>
Signed-off-by: Wadim Egorov <w.egorov@phytec.de>