media: ar0144: Only allow even numbered offsets
Selecting an odd numbered offset in either horizontal or vertical
direction would lead to a different bayer color order. Although this is
currently handled in code by forcing the corresponding bayer order
during set_fmt, this feature is not that useful. Thus remove it and
simplify the code. This results in only even numbered offsets to be
allowed when selecting the region of interest.

Signed-off-by: Stefan Riedmueller <s.riedmueller@phytec.de>
Signed-off-by: Teresa Remmet <t.remmet@phytec.de>
Signed-off-by: Wadim Egorov <w.egorov@phytec.de>