arm64: dts: ti: am62-phyboard-lyra: Always reserve R5 memory
The R5 co-processor is used to manage the AM62x SoC. Therefore, it
uses memory areas which are currently only resereved when the
co-processor overlay is activated (k3-am62-phyboard-lyra-rpmsg).

Always reserve the memory area in the phycore device tree file and
only activate the core in the overlay file.

Signed-off-by: Daniel Schultz <d.schultz@phytec.de>
Signed-off-by: Wadim Egorov <w.egorov@phytec.de>