media: ar0144: Put default sensor data in a separate structure
In preparation to support the ar0234 sensor with this driver add a
separate model_data structure which contains the default sensor data and
associate it with the devicetree compatible.

This way an additional sensor, like the ar0234, can easily be supported
by simply adding the new model_data for that sensor.

Signed-off-by: Stefan Riedmüller <s.riedmueller@phytec.de>
Tested-by: Dominik Haller <d.haller@phytec.de>
Signed-off-by: Wadim Egorov <w.egorov@phytec.de>