arm64: dts: am68x: phyboard-izar: Don't disable csi port nodes
There is actually no need to disable the csi port nodes since these are
subnodes of the cnds_csi2rx nodes and thus are only evaluated if that
node is enabled. Also there is no use case for enabling only specific
port nodes since the status of a port is determined by its remote
endpoint.

Signed-off-by: Stefan Riedmüller <s.riedmueller@phytec.de>
Tested-by: Dominik Haller <d.haller@phytec.de>
Signed-off-by: Wadim Egorov <w.egorov@phytec.de>