UPSTREAM: regmap-irq: Deprecate the not_fixed_stride flag
This flag is a bit of a hack and the same thing can be accomplished
using a custom ->get_irq_reg() callback. Add a warning to catch any
use of the flag.

Signed-off-by: Aidan MacDonald <aidanmacdonald.0x0@gmail.com>
Link: https://lore.kernel.org/r/20220623211420.918875-13-aidanmacdonald.0x0@gmail.com
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry-picked from commit 48e014ee9a61e8f4700987b82f7cb1dc3c89fa76 and adapt to current state of regmap)
Signed-off-by: Andrej Picej <andrej.picej@norik.com>
Signed-off-by: Wadim Egorov <w.egorov@phytec.de>