thermal: add support for two_stage thermal governor
Add support for 'two_stage' governor to explicitly enable and disable a
cooling device with only two stages as long as the temperature is still
above the trip_zone.

Existing 'step_wise' governor's trend-based throttling resulted in
frequent abrupt off and on toggling for a two-stage device and this
governor alleviates that behavior.

Signed-off-by: Russell Robinson <rrobinson@phytec.com>
Signed-off-by: khaleel <khaleel.s@phytec.in>
(fix commit msg, switch to THERMAL_GOVERNOR_DECLARE macro and rename to gov_two_stage.c)
Signed-off-by: Primoz Fiser <primoz.fiser@norik.com>
Reviewed-by: Daniel Schultz <d.schultz@phytec.de>
Signed-off-by: Wadim Egorov <w.egorov@phytec.de>