arm64: dts: ti: Added support for ADCs for AM64x
Added support for the ADCs on AM64x by adding the channel
configuration in order to fix the "missing channel configuration"
error.
The IO pins used by the ADC are exclusive to the ADC, so the channel
configuration was added to the phyCORE SoM device tree so the ADC
can be enabled for any board.

Signed-off-by: Nathan Morrisson <nmorrisson@phytec.com>
Signed-off-by: Wadim Egorov <w.egorov@phytec.de>